refactor: split lib out into multiple files
